requirements.txt
alembic==1.7.5
anyio==3.3.4
asgiref==3.4.1
asyncpg==0.25.0
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
decorator==5.1.0
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipython==7.29.0
jedi==0.18.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
toml==0.10.2
traitlets==5.1.1
typing_extensions==4.0.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
zipp==3.6.0
